tutorial/requirements.txt
python-dotenv==0.13.0
Django==3.0.6
djangorestframework==3.11.0
django-cors-headers==3.2.1
pygments==2.6.1  # for tutorial